requirements.txt
django==4.2
django-cors-headers==3.14.0
django-ninja==0.21.0
psycopg2

requests==2.28.1
beautifulsoup4==4.11.1
